(refactor) Tidy up messageCreate.ts
